@@ -4,13 +4,14 @@ class Order {
     public number: Number,
     public optionalAddress: string,
     public paymentOption: string,
-    public orderItems: OrderItem[] = []
+    public orderItems: OrderItem[] = [],
+    public id?: string
 
-  ) {}
+  ) { }
 }
 
 class OrderItem {
-  constructor(public quantity: number, public menuId: string) {}
+  constructor(public quantity: number, public menuId: string) { }
 }
 
 export { Order, OrderItem }
